backend/pyproject.toml
[tool.poetry]
name = "shop"
version = "0.1.0"
description = ""
authors = ["Iulia Zagornova <zagornowwa.julia@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.5"
django-model-utils = "^4.3.1"
pillow = "^10.0.1"
djangorestframework = "^3.14.0"
markdown = "^3.4.4"
django-filter = "^23.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
